Renames, regroups and cleanup
