+++ dbtmetabase/errors.py
@@ -1,0 +1,6 @@
+class ArgumentError(ValueError):
+    """Invalid argument supplied."""
+
+
+class MetabaseStateError(RuntimeError):
+    """Metabase in unexpected state."""
