+++ tests/__init__.py
@@ -1,9 +1,8 @@
 import logging
 
-from dbtmetabase._format import setup_logging
+from dbtmetabase.format import setup_logging
 
-from .test_core_exposures import *
-from .test_core_models import *
-from .test_manifest import *
+from .test_exposures import *
+from .test_models import *
 
 setup_logging(level=logging.DEBUG, path=None)
